Fixed an issue with remote miscast abilities

+++ src/content/game/LocalPlayerTurn.h
@@ -30,7 +30,7 @@ namespace padi::content {
     private:
         padi::UIContext *m_uiContext;
         std::vector<std::shared_ptr<sf::TcpSocket>> m_sockets;
-        int m_activeAbility = -1;
+        int8_t m_activeAbility = -1;
         bool m_hasCast = false;
     };
